Add files via upload
classes, objects

Design a ZProstokat class that describes the character rectangles on the console.

For each rectangle, specify:

width (int type),
height (int type),
the character needed to draw the edge of a rectangle on the console (char type),
the character needed to draw the inside of a rectangle on the console (char type).
The correct rectangle is the one in which the width and height should be positive natural numbers (i.e. from 1), while in the case where width (or height) = 1, the two listed signs should be the same.

Each rectangle (or in a special case, a square) should have its own unique number (starting with 1), given automatically when it is created (help: use the static components of the class). Rectangles and squares are numbered separately.